Cleanup split base and concretes
